requirements.txt
asgiref==3.7.2
astroid==2.15.8
black==23.11.0
click==8.1.7
dill==0.3.7
Django==5.0
django-filter==23.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dodgy==0.2.1
flake8==5.0.4
flake8-polyfill==1.0.2
gitdb==4.0.11
GitPython==3.1.40
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
pep8-naming==0.10.0
platformdirs==4.0.0
prospector==1.10.3
psycopg2==2.9.9
pycodestyle==2.9.1
pydocstyle==6.3.0
pyflakes==2.5.0
PyJWT==2.8.0
pylint==2.17.7
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pytz==2023.3.post1
PyYAML==6.0.1
requirements-detector==1.2.2
semver==3.0.2
setoptconf-tmp==0.3.1
setuptools==69.0.1
smmap==5.0.1
snowballstemmer==2.2.0
sqlparse==0.4.4
toml==0.10.2
tomlkit==0.12.3
wheel==0.41.3
wrapt==1.16.0
dependency_injector==4.41.0
